Move script to env - WIP - env vars not loading
