+++ .env_sample
@@ -8,3 +8,8 @@ TRANSIFEX_PROJECT=
 TRANSIFEX_SLUG=
 TRANSIFEX_API_TOKEN=
 LOCALES_PATH=
+GFW_ASSETS_PATH=
+GFW_API_KEY=
+TWITTER_WIDGET_API=
+FACEBOOK_WIDGET_API=
+GOOGLE_PLUS_ONE_WIDGET_API=
