+++ src/components/pages/landing/SocialFooter.js
@@ -1,30 +1,12 @@
 import React from 'react';
+import { SOCIAL_FOOTER_SCRIPT } from '../../..//constants/landing';
 
 class SocialFooter extends React.Component {
   componentDidMount(){
     this.script = document.createElement('script');
     this.script.type = 'text/javascript';
     this.script.async = true;
-    this.script.innerHTML = `
-      // Twitter
-      !function(d,s,id){var js,fjs=d.getElementsByTagName(s)[0];if(!d.getElementById(id)){js=d.createElement(s);js.id=id;js.src="//platform.twitter.com/widgets.js";fjs.parentNode.insertBefore(js,fjs);}}(document,"script","twitter-wjs");
-
-      // Facebook
-      (function(d, s, id) {
-        var js, fjs = d.getElementsByTagName(s)[0];
-        if (d.getElementById(id)) return;
-        js = d.createElement(s); js.id = id;
-        js.src = "//connect.facebook.net/en_US/all.js#xfbml=1&status=0";
-        fjs.parentNode.insertBefore(js, fjs);
-      }(document, 'script', 'facebook-jssdk'));
-
-      // Google Plus
-      (function() {
-        var po = document.createElement('script'); po.type = 'text/javascript'; po.async = true;
-        po.src = 'https://apis.google.com/js/plusone.js';
-        var s = document.getElementsByTagName('script')[0]; s.parentNode.insertBefore(po, s);
-      })();
-    `;
+    this.script.innerHTML = SOCIAL_FOOTER_SCRIPT;
     document.body.appendChild(this.script);
   }
